--- vmngclient/utils/timezones_enum.py
@@ -1,419 +1,0 @@
-from enum import Enum
-
-
-class Timezone(Enum):
-    Europe_Andorra: str = "Europe/Andorra"
-    Asia_Dubai: str = "Asia/Dubai"
-    Asia_Kabul: str = "Asia/Kabul"
-    America_Antigua: str = "America/Antigua"
-    America_Anguilla: str = "America/Anguilla"
-    Europe_Tirane: str = "Europe/Tirane"
-    Asia_Yerevan: str = "Asia/Yerevan"
-    Africa_Luanda: str = "Africa/Luanda"
-    Antarctica_McMurdo: str = "Antarctica/McMurdo"
-    Antarctica_Rothera: str = "Antarctica/Rothera"
-    Antarctica_Palmer: str = "Antarctica/Palmer"
-    Antarctica_Mawson: str = "Antarctica/Mawson"
-    Antarctica_Davis: str = "Antarctica/Davis"
-    Antarctica_Casey: str = "Antarctica/Casey"
-    Antarctica_Vostok: str = "Antarctica/Vostok"
-    Antarctica_DumontDUrville: str = "Antarctica/DumontDUrville"
-    Antarctica_Syowa: str = "Antarctica/Syowa"
-    America_Argentina_Buenos_Aires: str = "America/Argentina/Buenos_Aires"
-    America_Argentina_Cordoba: str = "America/Argentina/Cordoba"
-    America_Argentina_Salta: str = "America/Argentina/Salta"
-    America_Argentina_Jujuy: str = "America/Argentina/Jujuy"
-    America_Argentina_Tucuman: str = "America/Argentina/Tucuman"
-    America_Argentina_Catamarca: str = "America/Argentina/Catamarca"
-    America_Argentina_La_Rioja: str = "America/Argentina/La_Rioja"
-    America_Argentina_San_Juan: str = "America/Argentina/San_Juan"
-    America_Argentina_Mendoza: str = "America/Argentina/Mendoza"
-    America_Argentina_San_Luis: str = "America/Argentina/San_Luis"
-    America_Argentina_Rio_Gallegos: str = "America/Argentina/Rio_Gallegos"
-    America_Argentina_Ushuaia: str = "America/Argentina/Ushuaia"
-    Pacific_Pago_Pago: str = "Pacific/Pago_Pago"
-    Europe_Vienna: str = "Europe/Vienna"
-    Australia_Lord_Howe: str = "Australia/Lord_Howe"
-    Antarctica_Macquarie: str = "Antarctica/Macquarie"
-    Australia_Hobart: str = "Australia/Hobart"
-    Australia_Currie: str = "Australia/Currie"
-    Australia_Melbourne: str = "Australia/Melbourne"
-    Australia_Sydney: str = "Australia/Sydney"
-    Australia_Broken_Hill: str = "Australia/Broken_Hill"
-    Australia_Brisbane: str = "Australia/Brisbane"
-    Australia_Lindeman: str = "Australia/Lindeman"
-    Australia_Adelaide: str = "Australia/Adelaide"
-    Australia_Darwin: str = "Australia/Darwin"
-    Australia_Perth: str = "Australia/Perth"
-    Australia_Eucla: str = "Australia/Eucla"
-    America_Aruba: str = "America/Aruba"
-    Europe_Mariehamn: str = "Europe/Mariehamn"
-    Asia_Baku: str = "Asia/Baku"
-    Europe_Sarajevo: str = "Europe/Sarajevo"
-    America_Barbados: str = "America/Barbados"
-    Asia_Dhaka: str = "Asia/Dhaka"
-    Europe_Brussels: str = "Europe/Brussels"
-    Africa_Ouagadougou: str = "Africa/Ouagadougou"
-    Europe_Sofia: str = "Europe/Sofia"
-    Asia_Bahrain: str = "Asia/Bahrain"
-    Africa_Bujumbura: str = "Africa/Bujumbura"
-    Africa_Porto_Novo: str = "Africa/Porto-Novo"
-    America_St_Barthelemy: str = "America/St_Barthelemy"
-    Atlantic_Bermuda: str = "Atlantic/Bermuda"
-    Asia_Brunei: str = "Asia/Brunei"
-    America_La_Paz: str = "America/La_Paz"
-    America_Kralendijk: str = "America/Kralendijk"
-    America_Noronha: str = "America/Noronha"
-    America_Belem: str = "America/Belem"
-    America_Fortaleza: str = "America/Fortaleza"
-    America_Recife: str = "America/Recife"
-    America_Araguaina: str = "America/Araguaina"
-    America_Maceio: str = "America/Maceio"
-    America_Bahia: str = "America/Bahia"
-    America_Sao_Paulo: str = "America/Sao_Paulo"
-    America_Campo_Grande: str = "America/Campo_Grande"
-    America_Cuiaba: str = "America/Cuiaba"
-    America_Santarem: str = "America/Santarem"
-    America_Porto_Velho: str = "America/Porto_Velho"
-    America_Boa_Vista: str = "America/Boa_Vista"
-    America_Manaus: str = "America/Manaus"
-    America_Eirunepe: str = "America/Eirunepe"
-    America_Rio_Branco: str = "America/Rio_Branco"
-    America_Nassau: str = "America/Nassau"
-    Asia_Thimphu: str = "Asia/Thimphu"
-    Africa_Gaborone: str = "Africa/Gaborone"
-    Europe_Minsk: str = "Europe/Minsk"
-    America_Belize: str = "America/Belize"
-    America_St_Johns: str = "America/St_Johns"
-    America_Halifax: str = "America/Halifax"
-    America_Glace_Bay: str = "America/Glace_Bay"
-    America_Moncton: str = "America/Moncton"
-    America_Goose_Bay: str = "America/Goose_Bay"
-    America_Blanc_Sablon: str = "America/Blanc-Sablon"
-    America_Toronto: str = "America/Toronto"
-    America_Nipigon: str = "America/Nipigon"
-    America_Thunder_Bay: str = "America/Thunder_Bay"
-    America_Iqaluit: str = "America/Iqaluit"
-    America_Pangnirtung: str = "America/Pangnirtung"
-    America_Resolute: str = "America/Resolute"
-    America_Atikokan: str = "America/Atikokan"
-    America_Rankin_Inlet: str = "America/Rankin_Inlet"
-    America_Winnipeg: str = "America/Winnipeg"
-    America_Rainy_River: str = "America/Rainy_River"
-    America_Regina: str = "America/Regina"
-    America_Swift_Current: str = "America/Swift_Current"
-    America_Edmonton: str = "America/Edmonton"
-    America_Cambridge_Bay: str = "America/Cambridge_Bay"
-    America_Yellowknife: str = "America/Yellowknife"
-    America_Inuvik: str = "America/Inuvik"
-    America_Creston: str = "America/Creston"
-    America_Dawson_Creek: str = "America/Dawson_Creek"
-    America_Vancouver: str = "America/Vancouver"
-    America_Whitehorse: str = "America/Whitehorse"
-    America_Dawson: str = "America/Dawson"
-    Indian_Cocos: str = "Indian/Cocos"
-    Africa_Kinshasa: str = "Africa/Kinshasa"
-    Africa_Lubumbashi: str = "Africa/Lubumbashi"
-    Africa_Bangui: str = "Africa/Bangui"
-    Africa_Brazzaville: str = "Africa/Brazzaville"
-    Europe_Zurich: str = "Europe/Zurich"
-    Africa_Abidjan: str = "Africa/Abidjan"
-    Pacific_Rarotonga: str = "Pacific/Rarotonga"
-    America_Santiago: str = "America/Santiago"
-    Pacific_Easter: str = "Pacific/Easter"
-    Africa_Douala: str = "Africa/Douala"
-    Asia_Shanghai: str = "Asia/Shanghai"
-    Asia_Harbin: str = "Asia/Harbin"
-    Asia_Chongqing: str = "Asia/Chongqing"
-    Asia_Urumqi: str = "Asia/Urumqi"
-    Asia_Kashgar: str = "Asia/Kashgar"
-    America_Bogota: str = "America/Bogota"
-    America_Costa_Rica: str = "America/Costa_Rica"
-    America_Havana: str = "America/Havana"
-    Atlantic_Cape_Verde: str = "Atlantic/Cape_Verde"
-    America_Curacao: str = "America/Curacao"
-    Indian_Christmas: str = "Indian/Christmas"
-    Asia_Nicosia: str = "Asia/Nicosia"
-    Europe_Prague: str = "Europe/Prague"
-    Europe_Berlin: str = "Europe/Berlin"
-    Europe_Busingen: str = "Europe/Busingen"
-    Africa_Djibouti: str = "Africa/Djibouti"
-    Europe_Copenhagen: str = "Europe/Copenhagen"
-    America_Dominica: str = "America/Dominica"
-    America_Santo_Domingo: str = "America/Santo_Domingo"
-    Africa_Algiers: str = "Africa/Algiers"
-    America_Guayaquil: str = "America/Guayaquil"
-    Pacific_Galapagos: str = "Pacific/Galapagos"
-    Europe_Tallinn: str = "Europe/Tallinn"
-    Africa_Cairo: str = "Africa/Cairo"
-    Africa_El_Aaiun: str = "Africa/El_Aaiun"
-    Africa_Asmara: str = "Africa/Asmara"
-    Europe_Madrid: str = "Europe/Madrid"
-    Africa_Ceuta: str = "Africa/Ceuta"
-    Atlantic_Canary: str = "Atlantic/Canary"
-    Africa_Addis_Ababa: str = "Africa/Addis_Ababa"
-    Europe_Helsinki: str = "Europe/Helsinki"
-    Pacific_Fiji: str = "Pacific/Fiji"
-    Atlantic_Stanley: str = "Atlantic/Stanley"
-    Pacific_Chuuk: str = "Pacific/Chuuk"
-    Pacific_Pohnpei: str = "Pacific/Pohnpei"
-    Pacific_Kosrae: str = "Pacific/Kosrae"
-    Atlantic_Faroe: str = "Atlantic/Faroe"
-    Europe_Paris: str = "Europe/Paris"
-    Africa_Libreville: str = "Africa/Libreville"
-    Europe_London: str = "Europe/London"
-    America_Grenada: str = "America/Grenada"
-    Asia_Tbilisi: str = "Asia/Tbilisi"
-    America_Cayenne: str = "America/Cayenne"
-    Europe_Guernsey: str = "Europe/Guernsey"
-    Africa_Accra: str = "Africa/Accra"
-    Europe_Gibraltar: str = "Europe/Gibraltar"
-    America_Godthab: str = "America/Godthab"
-    America_Danmarkshavn: str = "America/Danmarkshavn"
-    America_Scoresbysund: str = "America/Scoresbysund"
-    America_Thule: str = "America/Thule"
-    Africa_Banjul: str = "Africa/Banjul"
-    Africa_Conakry: str = "Africa/Conakry"
-    America_Guadeloupe: str = "America/Guadeloupe"
-    Africa_Malabo: str = "Africa/Malabo"
-    Europe_Athens: str = "Europe/Athens"
-    Atlantic_South_Georgia: str = "Atlantic/South_Georgia"
-    America_Guatemala: str = "America/Guatemala"
-    Pacific_Guam: str = "Pacific/Guam"
-    Africa_Bissau: str = "Africa/Bissau"
-    America_Guyana: str = "America/Guyana"
-    Asia_Hong_Kong: str = "Asia/Hong_Kong"
-    America_Tegucigalpa: str = "America/Tegucigalpa"
-    Europe_Zagreb: str = "Europe/Zagreb"
-    America_Port_au_Prince: str = "America/Port-au-Prince"
-    Europe_Budapest: str = "Europe/Budapest"
-    Asia_Jakarta: str = "Asia/Jakarta"
-    Asia_Pontianak: str = "Asia/Pontianak"
-    Asia_Makassar: str = "Asia/Makassar"
-    Asia_Jayapura: str = "Asia/Jayapura"
-    Europe_Dublin: str = "Europe/Dublin"
-    Asia_Jerusalem: str = "Asia/Jerusalem"
-    Europe_Isle_of_Man: str = "Europe/Isle_of_Man"
-    Asia_Kolkata: str = "Asia/Kolkata"
-    Indian_Chagos: str = "Indian/Chagos"
-    Asia_Baghdad: str = "Asia/Baghdad"
-    Asia_Tehran: str = "Asia/Tehran"
-    Atlantic_Reykjavik: str = "Atlantic/Reykjavik"
-    Europe_Rome: str = "Europe/Rome"
-    Europe_Jersey: str = "Europe/Jersey"
-    America_Jamaica: str = "America/Jamaica"
-    Asia_Amman: str = "Asia/Amman"
-    Asia_Tokyo: str = "Asia/Tokyo"
-    Africa_Nairobi: str = "Africa/Nairobi"
-    Asia_Bishkek: str = "Asia/Bishkek"
-    Asia_Phnom_Penh: str = "Asia/Phnom_Penh"
-    Pacific_Tarawa: str = "Pacific/Tarawa"
-    Pacific_Enderbury: str = "Pacific/Enderbury"
-    Pacific_Kiritimati: str = "Pacific/Kiritimati"
-    Indian_Comoro: str = "Indian/Comoro"
-    America_St_Kitts: str = "America/St_Kitts"
-    Asia_Pyongyang: str = "Asia/Pyongyang"
-    Asia_Seoul: str = "Asia/Seoul"
-    Asia_Kuwait: str = "Asia/Kuwait"
-    America_Cayman: str = "America/Cayman"
-    Asia_Almaty: str = "Asia/Almaty"
-    Asia_Qyzylorda: str = "Asia/Qyzylorda"
-    Asia_Aqtobe: str = "Asia/Aqtobe"
-    Asia_Aqtau: str = "Asia/Aqtau"
-    Asia_Oral: str = "Asia/Oral"
-    Asia_Vientiane: str = "Asia/Vientiane"
-    Asia_Beirut: str = "Asia/Beirut"
-    America_St_Lucia: str = "America/St_Lucia"
-    Europe_Vaduz: str = "Europe/Vaduz"
-    Asia_Colombo: str = "Asia/Colombo"
-    Africa_Monrovia: str = "Africa/Monrovia"
-    Africa_Maseru: str = "Africa/Maseru"
-    Europe_Vilnius: str = "Europe/Vilnius"
-    Europe_Luxembourg: str = "Europe/Luxembourg"
-    Europe_Riga: str = "Europe/Riga"
-    Africa_Tripoli: str = "Africa/Tripoli"
-    Africa_Casablanca: str = "Africa/Casablanca"
-    Europe_Monaco: str = "Europe/Monaco"
-    Europe_Chisinau: str = "Europe/Chisinau"
-    Europe_Podgorica: str = "Europe/Podgorica"
-    America_Marigot: str = "America/Marigot"
-    Indian_Antananarivo: str = "Indian/Antananarivo"
-    Pacific_Majuro: str = "Pacific/Majuro"
-    Pacific_Kwajalein: str = "Pacific/Kwajalein"
-    Europe_Skopje: str = "Europe/Skopje"
-    Africa_Bamako: str = "Africa/Bamako"
-    Asia_Rangoon: str = "Asia/Rangoon"
-    Asia_Ulaanbaatar: str = "Asia/Ulaanbaatar"
-    Asia_Hovd: str = "Asia/Hovd"
-    Asia_Choibalsan: str = "Asia/Choibalsan"
-    Asia_Macau: str = "Asia/Macau"
-    Pacific_Saipan: str = "Pacific/Saipan"
-    America_Martinique: str = "America/Martinique"
-    Africa_Nouakchott: str = "Africa/Nouakchott"
-    America_Montserrat: str = "America/Montserrat"
-    Europe_Malta: str = "Europe/Malta"
-    Indian_Mauritius: str = "Indian/Mauritius"
-    Indian_Maldives: str = "Indian/Maldives"
-    Africa_Blantyre: str = "Africa/Blantyre"
-    America_Mexico_City: str = "America/Mexico_City"
-    America_Cancun: str = "America/Cancun"
-    America_Merida: str = "America/Merida"
-    America_Monterrey: str = "America/Monterrey"
-    America_Matamoros: str = "America/Matamoros"
-    America_Mazatlan: str = "America/Mazatlan"
-    America_Chihuahua: str = "America/Chihuahua"
-    America_Ojinaga: str = "America/Ojinaga"
-    America_Hermosillo: str = "America/Hermosillo"
-    America_Tijuana: str = "America/Tijuana"
-    America_Santa_Isabel: str = "America/Santa_Isabel"
-    America_Bahia_Banderas: str = "America/Bahia_Banderas"
-    Asia_Kuala_Lumpur: str = "Asia/Kuala_Lumpur"
-    Asia_Kuching: str = "Asia/Kuching"
-    Africa_Maputo: str = "Africa/Maputo"
-    Africa_Windhoek: str = "Africa/Windhoek"
-    Pacific_Noumea: str = "Pacific/Noumea"
-    Africa_Niamey: str = "Africa/Niamey"
-    Pacific_Norfolk: str = "Pacific/Norfolk"
-    Africa_Lagos: str = "Africa/Lagos"
-    America_Managua: str = "America/Managua"
-    Europe_Amsterdam: str = "Europe/Amsterdam"
-    Europe_Oslo: str = "Europe/Oslo"
-    Asia_Kathmandu: str = "Asia/Kathmandu"
-    Pacific_Nauru: str = "Pacific/Nauru"
-    Pacific_Niue: str = "Pacific/Niue"
-    Pacific_Auckland: str = "Pacific/Auckland"
-    Pacific_Chatham: str = "Pacific/Chatham"
-    Asia_Muscat: str = "Asia/Muscat"
-    America_Panama: str = "America/Panama"
-    America_Lima: str = "America/Lima"
-    Pacific_Tahiti: str = "Pacific/Tahiti"
-    Pacific_Marquesas: str = "Pacific/Marquesas"
-    Pacific_Gambier: str = "Pacific/Gambier"
-    Pacific_Port_Moresby: str = "Pacific/Port_Moresby"
-    Asia_Manila: str = "Asia/Manila"
-    Asia_Karachi: str = "Asia/Karachi"
-    Europe_Warsaw: str = "Europe/Warsaw"
-    America_Miquelon: str = "America/Miquelon"
-    Pacific_Pitcairn: str = "Pacific/Pitcairn"
-    America_Puerto_Rico: str = "America/Puerto_Rico"
-    Asia_Gaza: str = "Asia/Gaza"
-    Asia_Hebron: str = "Asia/Hebron"
-    Europe_Lisbon: str = "Europe/Lisbon"
-    Atlantic_Madeira: str = "Atlantic/Madeira"
-    Atlantic_Azores: str = "Atlantic/Azores"
-    Pacific_Palau: str = "Pacific/Palau"
-    America_Asuncion: str = "America/Asuncion"
-    Asia_Qatar: str = "Asia/Qatar"
-    Indian_Reunion: str = "Indian/Reunion"
-    Europe_Bucharest: str = "Europe/Bucharest"
-    Europe_Belgrade: str = "Europe/Belgrade"
-    Europe_Kaliningrad: str = "Europe/Kaliningrad"
-    Europe_Moscow: str = "Europe/Moscow"
-    Europe_Volgograd: str = "Europe/Volgograd"
-    Europe_Samara: str = "Europe/Samara"
-    Asia_Yekaterinburg: str = "Asia/Yekaterinburg"
-    Asia_Omsk: str = "Asia/Omsk"
-    Asia_Novosibirsk: str = "Asia/Novosibirsk"
-    Asia_Novokuznetsk: str = "Asia/Novokuznetsk"
-    Asia_Krasnoyarsk: str = "Asia/Krasnoyarsk"
-    Asia_Irkutsk: str = "Asia/Irkutsk"
-    Asia_Yakutsk: str = "Asia/Yakutsk"
-    Asia_Khandyga: str = "Asia/Khandyga"
-    Asia_Vladivostok: str = "Asia/Vladivostok"
-    Asia_Sakhalin: str = "Asia/Sakhalin"
-    Asia_Ust_Nera: str = "Asia/Ust-Nera"
-    Asia_Magadan: str = "Asia/Magadan"
-    Asia_Kamchatka: str = "Asia/Kamchatka"
-    Asia_Anadyr: str = "Asia/Anadyr"
-    Africa_Kigali: str = "Africa/Kigali"
-    Asia_Riyadh: str = "Asia/Riyadh"
-    Pacific_Guadalcanal: str = "Pacific/Guadalcanal"
-    Indian_Mahe: str = "Indian/Mahe"
-    Africa_Khartoum: str = "Africa/Khartoum"
-    Europe_Stockholm: str = "Europe/Stockholm"
-    Asia_Singapore: str = "Asia/Singapore"
-    Atlantic_St_Helena: str = "Atlantic/St_Helena"
-    Europe_Ljubljana: str = "Europe/Ljubljana"
-    Arctic_Longyearbyen: str = "Arctic/Longyearbyen"
-    Europe_Bratislava: str = "Europe/Bratislava"
-    Africa_Freetown: str = "Africa/Freetown"
-    Europe_San_Marino: str = "Europe/San_Marino"
-    Africa_Dakar: str = "Africa/Dakar"
-    Africa_Mogadishu: str = "Africa/Mogadishu"
-    America_Paramaribo: str = "America/Paramaribo"
-    Africa_Juba: str = "Africa/Juba"
-    Africa_Sao_Tome: str = "Africa/Sao_Tome"
-    America_El_Salvador: str = "America/El_Salvador"
-    America_Lower_Princes: str = "America/Lower_Princes"
-    Asia_Damascus: str = "Asia/Damascus"
-    Africa_Mbabane: str = "Africa/Mbabane"
-    America_Grand_Turk: str = "America/Grand_Turk"
-    Africa_Ndjamena: str = "Africa/Ndjamena"
-    Indian_Kerguelen: str = "Indian/Kerguelen"
-    Africa_Lome: str = "Africa/Lome"
-    Asia_Bangkok: str = "Asia/Bangkok"
-    Asia_Dushanbe: str = "Asia/Dushanbe"
-    Pacific_Fakaofo: str = "Pacific/Fakaofo"
-    Asia_Dili: str = "Asia/Dili"
-    Asia_Ashgabat: str = "Asia/Ashgabat"
-    Africa_Tunis: str = "Africa/Tunis"
-    Pacific_Tongatapu: str = "Pacific/Tongatapu"
-    Europe_Istanbul: str = "Europe/Istanbul"
-    America_Port_of_Spain: str = "America/Port_of_Spain"
-    Pacific_Funafuti: str = "Pacific/Funafuti"
-    Asia_Taipei: str = "Asia/Taipei"
-    Africa_Dar_es_Salaam: str = "Africa/Dar_es_Salaam"
-    Europe_Kiev: str = "Europe/Kiev"
-    Europe_Uzhgorod: str = "Europe/Uzhgorod"
-    Europe_Zaporozhye: str = "Europe/Zaporozhye"
-    Europe_Simferopol: str = "Europe/Simferopol"
-    Africa_Kampala: str = "Africa/Kampala"
-    Pacific_Johnston: str = "Pacific/Johnston"
-    Pacific_Midway: str = "Pacific/Midway"
-    Pacific_Wake: str = "Pacific/Wake"
-    America_New_York: str = "America/New_York"
-    America_Detroit: str = "America/Detroit"
-    America_Kentucky_Louisville: str = "America/Kentucky/Louisville"
-    America_Kentucky_Monticello: str = "America/Kentucky/Monticello"
-    America_Indiana_Indianapolis: str = "America/Indiana/Indianapolis"
-    America_Indiana_Vincennes: str = "America/Indiana/Vincennes"
-    America_Indiana_Winamac: str = "America/Indiana/Winamac"
-    America_Indiana_Marengo: str = "America/Indiana/Marengo"
-    America_Indiana_Petersburg: str = "America/Indiana/Petersburg"
-    America_Indiana_Vevay: str = "America/Indiana/Vevay"
-    America_Chicago: str = "America/Chicago"
-    America_Indiana_Tell_City: str = "America/Indiana/Tell_City"
-    America_Indiana_Knox: str = "America/Indiana/Knox"
-    America_Menominee: str = "America/Menominee"
-    America_North_Dakota_Center: str = "America/North_Dakota/Center"
-    America_North_Dakota_New_Salem: str = "America/North_Dakota/New_Salem"
-    America_North_Dakota_Beulah: str = "America/North_Dakota/Beulah"
-    America_Denver: str = "America/Denver"
-    America_Boise: str = "America/Boise"
-    America_Phoenix: str = "America/Phoenix"
-    America_Los_Angeles: str = "America/Los_Angeles"
-    America_Anchorage: str = "America/Anchorage"
-    America_Juneau: str = "America/Juneau"
-    America_Sitka: str = "America/Sitka"
-    America_Yakutat: str = "America/Yakutat"
-    America_Nome: str = "America/Nome"
-    America_Adak: str = "America/Adak"
-    America_Metlakatla: str = "America/Metlakatla"
-    Pacific_Honolulu: str = "Pacific/Honolulu"
-    America_Montevideo: str = "America/Montevideo"
-    Asia_Samarkand: str = "Asia/Samarkand"
-    Asia_Tashkent: str = "Asia/Tashkent"
-    Europe_Vatican: str = "Europe/Vatican"
-    America_St_Vincent: str = "America/St_Vincent"
-    America_Caracas: str = "America/Caracas"
-    America_Tortola: str = "America/Tortola"
-    America_St_Thomas: str = "America/St_Thomas"
-    Asia_Ho_Chi_Minh: str = "Asia/Ho_Chi_Minh"
-    Pacific_Efate: str = "Pacific/Efate"
-    Pacific_Wallis: str = "Pacific/Wallis"
-    Pacific_Apia: str = "Pacific/Apia"
-    Asia_Aden: str = "Asia/Aden"
-    Indian_Mayotte: str = "Indian/Mayotte"
-    Africa_Johannesburg: str = "Africa/Johannesburg"
-    Africa_Lusaka: str = "Africa/Lusaka"
-    Africa_Harare: str = "Africa/Harare"
